cpudetect.cpp: support compiling for Linux

+++ Code/Commando/WOLQuickMatch.cpp
@@ -290,7 +290,7 @@ bool WOLQuickMatch::SendClientInfo(void)
 	uint32_t ver = cNetwork::Get_Exe_Key();
 
 	// Get CPU speed
-	int speed = CPUDetectClass::Get_Processor_Speed();
+	uint64_t speed = CPUDetectClass::Get_Processor_Speed();
 
 	// Get amount of physical memory
 	MEMORYSTATUS memStatus;
@@ -330,7 +330,7 @@ bool WOLQuickMatch::SendClientInfo(void)
 	// Generate client information message
 	//-------------------------------------------------------------------------
 	WideStringClass clientMsg(256, true);
-	clientMsg.Format(U_CHAR("CINFO VER=%" PRIu32 " CPU=%lu MEM=%lu TPOINTS=%ld PLAYED=%lu PINGS=%S"),
+	clientMsg.Format(U_CHAR("CINFO VER=%" PRIu32 " CPU=%" PRIu64 " MEM=%lu TPOINTS=%ld PLAYED=%lu PINGS=%S"),
 		ver, speed, memory, tpoints, played, pseudoPings);
 
 	WWDEBUG_SAY(("WOLQuickMatch: '%S'\n", (const unichar_t*)clientMsg));
